fix: resolve log file hash calculation bug and refactor to shared utility
- Create shared hashPath() function in src/utils.ts
- Fix logger.ts hash calculation that always resulted in "0"
- Refactor daemon.ts to use shared utility instead of duplicate implementation
- Ensures log files use proper hashed directory names like PID and socket files
